move generate own sub package
